% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BigCity}
\alias{BigCity}
\title{Full Person-level Population Database}
\format{
A data frame with the following variables:
\describe{
  \item{HHID}{The identifier of the household. It corresponds to an
    alphanumeric sequence (four letters and five digits).}
  \item{PersonID}{The identifier of the person within the household.
    Note: it is not a unique identifier for the whole population.}
  \item{Stratum}{Households are located in geographic strata. There are
    119 strata across the city.}
  \item{PSU}{Households are clustered in cartographic segments defined as
    primary sampling units (PSU). There are 1,664 PSU nested within strata.}
  \item{Zone}{Segments within strata can be located in urban or rural
    areas across the city.}
  \item{Sex}{Sex of the person.}
  \item{Age}{Age of the person.}
  \item{Income}{Per capita monthly income.}
  \item{Expenditure}{Per capita monthly expenditure.}
  \item{Employment}{A person's employment status.}
  \item{Poverty}{Indicates whether the person is poor or not, based on
    income.}
  \item{MaritalST}{Marital status of the person.}
}
}
\usage{
data(BigCity)
}
\description{
A data frame corresponding to some socioeconomic variables from 150,266
people of a city in a particular year.
}
\examples{
data(BigCity)
attach(BigCity)
estima <- data.frame(Income, Expenditure)
# The population totals
colSums(estima)
# Some parameters of interest
table(Poverty, Zone)
xtabs(Income ~ Poverty + Zone)
# Correlations among characteristics of interest
cor(estima)
# Some useful histograms
hist(Income)
hist(Expenditure)
# Some useful plots
boxplot(Income ~ Poverty)
barplot(table(Employment))
pie(table(MaritalST))
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{Lucy}}, \code{\link{BigLucy}}
}
\author{
Hugo Andres Gutierrez Rojas \email{hagutierrezro@gmail.com}
}
\keyword{datasets}
