% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BigLucy}
\alias{BigLucy}
\title{Full Business Population Database (Extended)}
\format{
A data frame with the following variables:
\describe{
  \item{ID}{The identifier of the company. It corresponds to an
    alphanumeric sequence (two letters and three digits).}
  \item{Ubication}{The address of the principal office of the company
    in the city.}
  \item{Level}{The industrial companies are discriminated according to the
    taxes declared. There are small, medium and big companies.}
  \item{Zone}{The country is divided by counties. A company belongs to a
    particular zone according to its cartographic location.}
  \item{Income}{The total amount of a company's earnings in the previous
    fiscal year.}
  \item{Employees}{The total number of persons working for the company in
    the previous fiscal year.}
  \item{Taxes}{The total amount of a company's income tax.}
  \item{SPAM}{Indicates if the company uses the Internet and webmail
    options to make self-propaganda.}
  \item{ISO}{Indicates if the company is certified by the International
    Organization for Standardization.}
  \item{Years}{The age of the company.}
  \item{Segments}{Cartographic segments by county. A segment comprises
    on average 10 companies located close to each other.}
}
}
\usage{
data(BigLucy)
}
\description{
A data frame corresponding to some financial variables of 85,396 industrial
companies of a city in a particular fiscal year.
}
\examples{
data(BigLucy)
attach(BigLucy)
# The variables of interest are: Income, Employees and Taxes
estima <- data.frame(Income, Employees, Taxes)
# The population totals
colSums(estima)
# Some parameters of interest
table(SPAM, Level)
xtabs(Income ~ Level + SPAM)
# Correlations among characteristics of interest
cor(estima)
# Some useful histograms
hist(Income)
hist(Taxes)
hist(Employees)
# Some useful plots
boxplot(Income ~ Level)
barplot(table(Level))
pie(table(SPAM))
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{Lucy}}, \code{\link{BigCity}}
}
\author{
Hugo Andres Gutierrez Rojas \email{hagutierrezro@gmail.com}
}
\keyword{datasets}
