% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.piPS.r
\name{E.piPS}
\alias{E.piPS}
\title{Estimation of the Population Total under Probability-Proportional-to-Size Sampling Without Replacement}
\usage{
E.piPS(y, Pik)
}
\arguments{
\item{y}{Vector, matrix or data frame containing the values of the
variables of interest for every unit in the selected sample.}

\item{Pik}{Vector of first-order inclusion probabilities for each
unit in the sample.}
}
\value{
A matrix with four rows and one column per variable of interest:
\itemize{
  \item \code{Estimation}: Estimated population total.
  \item \code{Standard Error}: Estimated standard error of the total.
  \item \code{CVE}: Estimated coefficient of variation (in percentage).
  \item \code{DEFF}: Design effect with respect to simple random sampling.
}
}
\description{
Computes the Horvitz-Thompson estimator of the population total under a
without-replacement probability proportional to size (piPS) sampling design.
The variance is estimated using the Horvitz-Thompson variance approximation
based on first-order inclusion probabilities.
}
\details{
When all inclusion probabilities are equal (i.e. \code{sum(Pik) == n}),
the variance is set to zero, reflecting an equal-probability design.
}
\examples{
data('Lucy')
attach(Lucy)
N <- nrow(Lucy)
n <- 400
x <- Employees
res <- S.piPS(n, x)
sam <- res[, 1]
Pik <- res[, 2]
y   <- data.frame(Income = Income[sam], Taxes = Taxes[sam])
E.piPS(y, Pik)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{S.piPS}}, \code{\link{PikPPS}}, \code{\link{E.PO}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
