% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.PPS.r
\name{E.PPS}
\alias{E.PPS}
\title{Estimation of the Population Total under PPS With-Replacement Sampling}
\usage{
E.PPS(y, pk)
}
\arguments{
\item{y}{Vector, matrix or data frame containing the values of the
variables of interest for every selected unit (with possible repetitions).}

\item{pk}{Vector of selection probabilities for each draw in the sample.}
}
\value{
A matrix with four rows and one column per variable of interest:
\itemize{
  \item \code{Estimation}: Estimated population total.
  \item \code{Standard Error}: Estimated standard error of the total.
  \item \code{CVE}: Estimated coefficient of variation (in percentage).
  \item \code{DEFF}: Design effect with respect to simple random sampling.
}
}
\description{
Computes the Hansen-Hurwitz estimator of the population total under a
probability proportional to size with-replacement (PPS-WR) sampling design.
}
\details{
The Hansen-Hurwitz estimator is \eqn{\hat{t} = (1/m)\sum_{i=1}^m y_i/p_i},
where \eqn{p_i} is the selection probability of the \eqn{i}-th draw and
\eqn{m} is the number of draws.
}
\examples{
data('Lucy')
attach(Lucy)
m   <- 400
res <- S.PPS(m, Employees)
sam <- res[, 1]
pk  <- res[, 2]
y   <- data.frame(Income = Income[sam], Expenditure = Expenditure[sam])
E.PPS(y, pk)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{S.PPS}}, \code{\link{HH}}, \code{\link{E.piPS}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
