% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.STpiPS.R
\name{E.STpiPS}
\alias{E.STpiPS}
\title{Estimation of the Population Total under Stratified Probability Proportional to Size Sampling Without Replacement}
\usage{
E.STpiPS(y, Pik, S)
}
\arguments{
\item{y}{Vector, matrix or data frame of variables of interest.}

\item{Pik}{Vector of first-order inclusion probabilities for each unit
in the sample.}

\item{S}{Vector identifying the stratum membership of each unit in the sample.}
}
\value{
A matrix with four rows and one column per variable of interest:
\itemize{
  \item \code{Estimation}: Estimated population total.
  \item \code{Standard Error}: Estimated standard error.
  \item \code{CVE}: Estimated coefficient of variation (in percentage).
  \item \code{DEFF}: Design effect with respect to simple random sampling.
}
}
\description{
Computes the Horvitz-Thompson estimator of the population total under a
stratified without-replacement probability proportional to size (piPS)
sampling design.
}
\examples{
# Uses the Lucy data to draw a stratified random sample
# according to a piPS design in each stratum
data(Lucy)
attach(Lucy)
N1 <- summary(Level)[[1]]
N2 <- summary(Level)[[2]]
N3 <- summary(Level)[[3]]
nh <- c(N1, 100, 200)
S  <- Level
x  <- Employees
res <- S.STpiPS(S, x, nh)
sam <- res[, 1]
pik <- res[, 2]
data <- Lucy[sam, ]
attach(data)
estima <- data.frame(Income, Employees, Taxes)
E.STpiPS(estima, pik, Level)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{S.STpiPS}}, \code{\link{E.piPS}}, \code{\link{E.STSI}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
