% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.STPPS.r
\name{E.STPPS}
\alias{E.STPPS}
\title{Estimation of the Population Total under Stratified Probability Proportional to Size Sampling With Replacement}
\usage{
E.STPPS(y, pk, mh, S)
}
\arguments{
\item{y}{Vector, matrix or data frame of variables of interest.}

\item{pk}{Vector of selection probabilities for each draw in the sample.}

\item{mh}{Integer vector with the number of draws within each stratum.}

\item{S}{Vector identifying the stratum membership of each unit in the sample.}
}
\value{
A matrix with four rows and one column per variable of interest:
\itemize{
  \item \code{Estimation}: Estimated population total.
  \item \code{Standard Error}: Estimated standard error.
  \item \code{CVE}: Estimated coefficient of variation (in percentage).
  \item \code{DEFF}: Design effect with respect to simple random sampling.
}
}
\description{
Computes the Hansen-Hurwitz estimator of the population total under a
stratified PPS with-replacement (STPPS) sampling design.
}
\examples{
# Uses the Lucy data to draw a stratified random sample
# according to a PPS design in each stratum
data(Lucy)
attach(Lucy)
m1 <- 83; m2 <- 100; m3 <- 200
mh <- c(m1, m2, m3)
res <- S.STPPS(Level, Income, mh)
sam <- res[, 1]
pk  <- res[, 2]
data <- Lucy[sam, ]
attach(data)
estima <- data.frame(Income, Employees, Taxes)
E.STPPS(estima, pk, mh, Level)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{S.STPPS}}, \code{\link{E.PPS}}, \code{\link{E.STpiPS}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
