% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.WR.r
\name{E.WR}
\alias{E.WR}
\title{Estimation of the Population Total under Simple Random Sampling With
Replacement}
\usage{
E.WR(N, m, y)
}
\arguments{
\item{N}{Population size.}

\item{m}{Number of draws (sample size with replacement).}

\item{y}{Vector, matrix or data frame containing the values of the
variables of interest for every draw in the sample (repetitions allowed).}
}
\value{
A matrix with four rows and one column per variable of interest:
\itemize{
  \item \code{Estimation}: Estimated population total.
  \item \code{Standard Error}: Estimated standard error of the total.
  \item \code{CVE}: Estimated coefficient of variation (in percentage).
  \item \code{DEFF}: Design effect with respect to simple random sampling
    without replacement.
}
}
\description{
Computes the Hansen-Hurwitz estimator of the population total under a
simple random sampling with replacement (WR) design.
}
\details{
Under simple random sampling with replacement with \code{m} draws, the
Hansen-Hurwitz estimator is \eqn{\hat{t} = (N/m)\sum_{i=1}^m y_i}.
}
\examples{
data('Lucy')
attach(Lucy)
N   <- nrow(Lucy)
m   <- 400
sam <- S.WR(N, m)
y   <- data.frame(Income = Income[sam], Expenditure = Expenditure[sam])
E.WR(N, m, y)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{S.WR}}, \code{\link{E.SI}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
