% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lucy}
\alias{Lucy}
\title{Full Business Population Database}
\format{
A data frame with 2,396 rows and 8 variables:
\describe{
  \item{ID}{The identifier of the company. It corresponds to an
    alphanumeric sequence (two letters and ten digits).}
  \item{Ubication}{The address of the principal office of the company.}
  \item{Level}{The size level of the company discriminated according to
    the income declared. There are small, medium and big companies.}
  \item{Zone}{The country is divided by counties. A company belongs to a
    particular zone according to its cartographic location.}
  \item{Income}{The total amount of a company's earnings in the previous
    fiscal year.}
  \item{Employees}{The total number of persons working for the company in
    the previous fiscal year.}
  \item{Taxes}{The total amount of a company's income tax.}
  \item{SPAM}{Indicates if the company uses the Internet and webmail
    options to make self-propaganda.}
}
}
\usage{
data(Lucy)
}
\description{
A data frame containing socioeconomic and financial variables from a
population of 2,396 business units, used throughout the package to
illustrate survey sampling designs and estimators.
}
\examples{
data(Lucy)
attach(Lucy)
# The variables of interest are: Income, Employees and Taxes
estima <- data.frame(Income, Employees, Taxes)
# The population totals
colSums(estima)
# Some parameters of interest
table(SPAM, Level)
xtabs(Income ~ Level + SPAM)
# Correlations among characteristics of interest
cor(estima)
# Some useful histograms
hist(Income)
hist(Taxes)
hist(Employees)
# Some useful plots
boxplot(Income ~ Level)
barplot(table(Level))
pie(table(SPAM))
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{BigLucy}}, \code{\link{BigCity}}
}
\author{
Hugo Andres Gutierrez Rojas \email{hagutierrezro@gmail.com}
}
\keyword{datasets}
