% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderWR.r
\name{OrderWR}
\alias{OrderWR}
\title{Ordered With-Replacement Sampling Support}
\usage{
OrderWR(N, m, ID = FALSE)
}
\arguments{
\item{N}{Population size.}

\item{m}{Number of draws.}

\item{ID}{Optional vector of population labels of length \code{N}.
If provided, labels are substituted for integer indices in the output.
If \code{FALSE} (default), integer indices are returned.}
}
\value{
A matrix with \code{N^m} rows and \code{m} columns, where each row is one
ordered sequence of draws. If \code{ID} is provided, population labels are
substituted for indices.
}
\description{
Enumerates all ordered sequences of \code{m} draws from a population of
size \code{N} with replacement. Unlike \code{\link{SupportWR}}, this
function considers order, so sequences that differ only in draw order are
treated as distinct outcomes.
}
\details{
The total number of ordered with-replacement sequences of size \code{m}
from \code{N} units is \eqn{N^m}. This grows rapidly and the function
should only be used for small \code{N} and \code{m}.
}
\examples{
U <- c("Yves", "Ken", "Erik", "Sharon", "Leslie")
N <- length(U)
# Five possible ordered samples of size m=1
OrderWR(N, 1)
OrderWR(N, 1, ID = U)
# 25 possible ordered samples of size m=2
OrderWR(N, 2)
OrderWR(N, 2, ID = U)
# Note: ordered samples differ from unordered (SupportWR)
OrderWR(N, 2)
SupportWR(N, 2)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{SupportWR}}, \code{\link{IkWR}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
