% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.piPS.r
\name{S.piPS}
\alias{S.piPS}
\title{Probability Proportional to Size Without-Replacement Sampling (piPS)}
\usage{
S.piPS(n, x, e = runif(length(x)))
}
\arguments{
\item{n}{Sample size.}

\item{x}{Vector of length \code{N} with positive auxiliary size values.}

\item{e}{Optional vector of \code{N} uniform random variates in \code{(0,1)}.
If omitted, \code{runif(N)} is used.}
}
\value{
A matrix with \code{n} rows and two columns:
\itemize{
  \item Column 1: population indices of the selected units.
  \item Column 2: first-order inclusion probabilities of the selected units.
}
}
\description{
Draws a without-replacement sample of size \code{n} using a sequential
algorithm that produces inclusion probabilities proportional to an
auxiliary size variable \code{x}.
}
\examples{
############
## Example 1
############
U <- c("Yves", "Ken", "Erik", "Sharon", "Leslie")
x <- c(52, 60, 75, 100, 50)
# Draws a piPS sample without replacement of size n=3
res <- S.piPS(3, x)
res
sam <- res[, 1]
U[sam]
############
## Example 2
############
# Uses the Lucy data
data(Lucy)
attach(Lucy)
res <- S.piPS(400, Income)
sam <- res[, 1]
Pik.s <- res[, 2]
data <- Lucy[sam, ]
dim(data)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{E.piPS}}, \code{\link{PikPPS}}, \code{\link{S.STPPS}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
