% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.PO.r
\name{S.PO}
\alias{S.PO}
\title{Poisson Sampling}
\usage{
S.PO(N, Pik)
}
\arguments{
\item{N}{Population size.}

\item{Pik}{Vector of length \code{N} containing the first-order inclusion
probability for each unit in the population. Values must be in \code{(0, 1]}.}
}
\value{
A vector of length \code{N} where selected units contain their population
index and non-selected units contain \code{0}.
}
\description{
Draws a Poisson sample from a finite population of size \code{N}.
Each unit \eqn{k} is independently selected with its own inclusion
probability \eqn{\pi_k}.
}
\details{
Poisson sampling is a generalisation of Bernoulli sampling that allows
unequal inclusion probabilities. The sample size is random. To extract
the selected indices, use \code{sam[sam != 0]}.
}
\examples{
data('Lucy')
attach(Lucy)
N   <- nrow(Lucy)
n   <- 400
Pik <- PikPPS(n, Employees)
sam <- S.PO(N, Pik)
sam <- sam[sam != 0]
y   <- data.frame(Income = Income[sam], Expenditure = Expenditure[sam])
E.PO(y, Pik[sam])
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{E.PO}}, \code{\link{PikPPS}}, \code{\link{S.piPS}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
