% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wk.r
\name{Wk}
\alias{Wk}
\title{GREG Generalised Weights}
\usage{
Wk(x, tx, Pik, ck, b0 = FALSE)
}
\arguments{
\item{x}{Vector or matrix of auxiliary variables observed in the sample.}

\item{tx}{Vector of known population totals of the auxiliary variables.}

\item{Pik}{Vector of first-order inclusion probabilities for each unit
in the sample.}

\item{ck}{Vector of variance-stabilising constants. Typically \code{ck = 1}
(homoscedastic) or \code{ck = x} (heteroscedastic).}

\item{b0}{Logical. If \code{TRUE}, an intercept column is prepended to
\code{x}. Default is \code{FALSE}.}
}
\value{
A numeric vector of length \code{n} with the GREG weight for each unit
in the sample.
}
\description{
Computes the generalised regression (GREG) weights for each unit in the
sample. These weights incorporate both the sampling design weights and a
calibration adjustment based on known population totals of auxiliary
variables.
}
\details{
The GREG weight for unit \eqn{k} is:
\deqn{w_k = \frac{1}{\pi_k} + \mathbf{x}_k^T
\left(\sum_s \frac{v_k \mathbf{x}_k \mathbf{x}_k^T}{\pi_k}\right)^{-1}
(\mathbf{t}_x - \hat{\mathbf{t}}_{x,\pi})}
where \eqn{v_k = 1/(\pi_k c_k)} and \eqn{c_k} is a variance-stabilising
constant. The GREG estimator is then \eqn{\hat{t}_{GREG} = \sum_s w_k y_k}.
}
\examples{
data('Lucy')
attach(Lucy)
N   <- nrow(Lucy)
n   <- 400
sam <- S.SI(N, n)
Pik <- rep(n/N, n)
x   <- as.matrix(Employees[sam])
tx  <- sum(Employees)
ck  <- rep(1, n)
wk  <- Wk(x, tx, Pik, ck)
# Check calibration: weighted sum of x equals tx
sum(wk * x)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{GREG.SI}}, \code{\link{E.Beta}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
