% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.PPS.r
\name{S.PPS}
\alias{S.PPS}
\title{Probability Proportional to Size With-Replacement Sampling}
\usage{
S.PPS(m, x)
}
\arguments{
\item{m}{Number of draws (sample size with replacement).}

\item{x}{Vector of length \code{N} containing positive auxiliary size
values for each unit in the population.}
}
\value{
A matrix with \code{m} rows and two columns:
\itemize{
  \item Column 1 (\code{sam}): population indices of the selected units.
  \item Column 2 (\code{pk}): selection probability of each draw.
}
}
\description{
Draws a with-replacement sample of size \code{m} from a finite population
using probabilities proportional to an auxiliary size variable \code{x}.
}
\details{
At each draw, unit \eqn{k} is selected with probability
\eqn{p_k = x_k / \sum x}. Since sampling is with replacement, the same
unit may appear more than once. Use \code{\link{E.PPS}} or \code{\link{HH}}
to estimate population totals from this sample.
}
\examples{
############
## Example 1
############
# Vector U contains the label of a population of size N=5
U <- c("Yves", "Ken", "Erik", "Sharon", "Leslie")
# The auxiliary information
x <- c(52, 60, 75, 100, 50)
# Draws a PPS sample with replacement of size m=3
res <- S.PPS(3,x)
sam <- res[,1]
# The selected sample is
U[sam]

############
## Example 2
############
# Uses the Lucy data to draw a random sample according to a 
# PPS with replacement design
data(Lucy)
attach(Lucy)
# The selection probability of each unit is proportional to the variable Income
m <- 400
res<-S.PPS(400,Income)
# The selected sample
sam <- res[,1]
# The information about the units in the sample is stored in an object called data
data <- Lucy[sam,]
data
dim(data)
}
\references{
Sarndal, C-E. and Swensson, B. and Wretman, J. (1992),
\emph{Model Assisted Survey Sampling}. Springer.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{E.PPS}}, \code{\link{HH}}, \code{\link{S.piPS}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
