% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.SI.r
\name{S.SI}
\alias{S.SI}
\title{Simple Random Sampling Without Replacement}
\usage{
S.SI(N, n, e = runif(N))
}
\arguments{
\item{N}{Population size.}

\item{n}{Sample size. Must satisfy \code{n <= N}.}

\item{e}{Optional vector of \code{N} uniform random variates in \code{(0,1)}.
If omitted, \code{runif(N)} is used. Useful for reproducibility or
coordinated sampling.}
}
\value{
A vector of length \code{N} where selected units contain their population
index and non-selected units contain \code{0}.
}
\description{
Draws a simple random sample of size \code{n} without replacement from a
finite population of size \code{N} using the sequential algorithm of
Fan, Muller and Rezucha (1962).
}
\details{
The sequential algorithm selects units one at a time by comparing a uniform
random variate with the conditional inclusion probability at each step,
ensuring exactly \code{n} units are selected. To extract the selected
indices, filter out the zeros: \code{sam[sam != 0]}.
}
\examples{
############
## Example 1
############
# Vector U contains the label of a population of size N=5
U <- c("Yves", "Ken", "Erik", "Sharon", "Leslie")
# Fixes the random numbers in order to select a sample
e <- c(0.4938, 0.7044, 0.4585, 0.6747, 0.0640)
# Draws a simple random sample without replacement of size n=3
sam <- S.SI(5, 3, e)
sam
# The selected sample is
U[sam]
############
## Example 2
############
# Uses the Lucy data to draw a random sample according to a SI design
data(Lucy)
attach(Lucy)
N <- dim(Lucy)[1]
n <- 400
sam <- S.SI(N, n)
# The information about the units in the sample
data <- Lucy[sam, ]
dim(data)
}
\references{
Fan, C.T., Muller, M.E. and Rezucha, I. (1962). Development of sampling
plans by using sequential (item by item) selection techniques and digital
computers. \emph{Journal of the American Statistical Association},
57(298), 387-402.\cr
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas
y estimacion de parametros}. Editorial Universidad Santo Tomas.
}
\seealso{
\code{\link{E.SI}}, \code{\link{S.STSI}}, \code{\link{S.SY}}
}
\author{
Hugo Andres Gutierrez Rojas <hagutierrezro at gmail.com>
}
